pyproject.toml
[project]
name = "01-getting-started"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "chainlit>=2.5.5",
    "openai-agents>=0.2.2",
    "python-dotenv>=1.1.0",
    "streamlit>=1.45.1",
]
